Fix #329 Use a different settings.py for each environment
