+++ tcf_core/settings/base.py
@@ -1,6 +1,6 @@
-# pylint: disable=line-too-long,import-error,fixme
+# pylint: disable=fixme
 """
-Django settings for tcf_core project.
+Base Django settings for tcf_core project.
 
 Generated by 'django-admin startproject' using Django 3.0.4.
 
@@ -12,25 +12,27 @@ https://docs.djangoproject.com/en/3.0/ref/settings/
 """
 
 import os
+
 from django.urls import reverse_lazy
 from django.contrib.messages import constants as messages
 import environ
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 # Django-environ library imports .env settings
-env = environ.Env(DEBUG=(bool, False))
-env_file = os.path.join(BASE_DIR, ".env")
+env = environ.Env(
+    # set casting, default value
+    DEBUG=(bool, False),
+)
+env_file = os.path.join(BASE_DIR, '.env')
 environ.Env.read_env(env_file)
-
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = env.str('SECRET_KEY')
 
 # SECURITY WARNING: don't run with debug turned on in production!
-# DEBUG = int(env.str('DEBUG', 0)) == 1      # Why can't debug be a bool?
-DEBUG = env.str('DEBUG', default=False)
+DEBUG = env.bool('DEBUG')  # default value set on the top
 
 ALLOWED_HOSTS = ['localhost', '.ngrok.io', '127.0.0.1']
 
@@ -47,7 +49,7 @@ INSTALLED_APPS = [
     'cachalot',  # TODO: add Redis?
     'rest_framework',
     'django_filters',
-    'tcf_website'
+    'tcf_website',
 ]
 
 MIDDLEWARE = [
@@ -121,8 +123,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/3.0/howto/static-files/
 
 STATIC_URL = '/static/'
-STATIC_ROOT = os.path.join(BASE_DIR, "static/")
-# STATIC_URL = env.str('STATIC_URL', default='static/')
+STATIC_ROOT = os.path.join(BASE_DIR, 'static/')
 
 
 # Database
@@ -130,11 +131,12 @@ STATIC_ROOT = os.path.join(BASE_DIR, "static/")
 
 DATABASES = {
     'default': {
-        'NAME': 'tcf_db',
+        'NAME': env.str('DB_NAME'),
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
-        'USER': 'tcf_django',
-        'PASSWORD': 's3kr1t',
-        'HOST': 'localhost' if os.environ.get('CI') else 'tcf_db',
+        'USER': env.str('DB_USER'),
+        'PASSWORD': env.str('DB_PASSWORD'),
+        'HOST': env.str('DB_HOST'),
+        'PORT': env.int('DB_PORT')
     },
     'legacy': {
         'ENGINE': 'django.db.backends.sqlite3',
@@ -143,24 +145,18 @@ DATABASES = {
 }
 
 
-# python-social-auth settings.
+# social-auth-app-django settings.
 
 AUTHENTICATION_BACKENDS = (
     'social_core.backends.google.GoogleOAuth2',
     'django.contrib.auth.backends.ModelBackend',
 )
-
+# TODO: Look into options like SOCIAL_AUTH_LOGIN_ERROR_URL or LOGIN_ERROR_URL
 SOCIAL_AUTH_GOOGLE_OAUTH2_KEY = env.str('SOCIAL_AUTH_GOOGLE_OAUTH2_KEY')
 SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET = env.str('SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET')
 SOCIAL_AUTH_GOOGLE_OAUTH2_WHITELISTED_DOMAINS = ['virginia.edu']
 SOCIAL_AUTH_LOGIN_REDIRECT_URL = reverse_lazy('browse')
-# SOCIAL_AUTH_LOGIN_ERROR_URL = '/'
-# LOGIN_ERROR_URL = '/'
 SOCIAL_AUTH_RAISE_EXCEPTIONS = False
-
-SOCIAL_AUTH_USER_MODEL = 'tcf_website.User'
-AUTH_USER_MODEL = 'tcf_website.User'
-
 SOCIAL_AUTH_PIPELINE = (
     'social_core.pipeline.social_auth.social_details',
     'social_core.pipeline.social_auth.social_uid',
@@ -174,80 +170,38 @@ SOCIAL_AUTH_PIPELINE = (
     'social_core.pipeline.social_auth.load_extra_data',
     'social_core.pipeline.user.user_details',
 )
+SOCIAL_AUTH_USER_MODEL = 'tcf_website.User'
+
+AUTH_USER_MODEL = 'tcf_website.User'
 
 # Read-only access to Elastic
 ES_PUBLIC_API_KEY = env.str('ES_PUBLIC_API_KEY')
 ES_COURSE_SEARCH_ENDPOINT = env.str('ES_COURSE_SEARCH_ENDPOINT')
 ES_INSTRUCTOR_SEARCH_ENDPOINT = env.str('ES_INSTRUCTOR_SEARCH_ENDPOINT')
 
-# PROD SETTINGS
-if not DEBUG:
+# Read-write access to Elastic
+ES_COURSE_DOCUMENTS_ENDPOINT = env.str('ES_COURSE_DOCUMENTS_ENDPOINT')
+ES_INSTRUCTOR_DOCUMENTS_ENDPOINT = env.str(
+    'ES_INSTRUCTOR_DOCUMENTS_ENDPOINT')
+ES_PRIVATE_API_KEY = env.str('ES_PRIVATE_API_KEY')
 
-    # Heroku configuration.
-    if env.bool("HEROKU", default=False):
-        import django_heroku
-        django_heroku.settings(locals())
-
-    # Gather information from environment variables.
-
-    # HOSTNAME = env.str('HOSTNAME')
-    # PUBLIC_IPV4 = env.str('PUBLIC_IPV4')
-
-    # SECURITY WARNING: App Engine's security features ensure that it is safe to
-    # have ALLOWED_HOSTS = ['*'] when the app is deployed. If you deploy a Django
-    # app not on App Engine, make sure to set an appropriate host here.
-    # See https://docs.djangoproject.com/en/1.10/ref/settings/ (from GCP
-    # documentation)
-    ALLOWED_HOSTS = ['*', 'thecourseforum.com']
-
-    # if HOSTNAME:
-    #     ALLOWED_HOSTS.append(HOSTNAME)
-    # if PUBLIC_IPV4:
-    #     ALLOWED_HOSTS.append(PUBLIC_IPV4)
-
-    # Read-write access to Elastic
-    ES_COURSE_DOCUMENTS_ENDPOINT = env.str('ES_COURSE_DOCUMENTS_ENDPOINT')
-    ES_INSTRUCTOR_DOCUMENTS_ENDPOINT = env.str(
-        'ES_INSTRUCTOR_DOCUMENTS_ENDPOINT')
-    ES_PRIVATE_API_KEY = env.str('ES_PRIVATE_API_KEY')
-
-    DATABASES['default'] = {
-        'NAME': env.str('DB_NAME'),
-        'ENGINE': 'django.db.backends.postgresql_psycopg2',
-        'USER': env.str('DB_USER'),
-        'PASSWORD': env.str('DB_PASSWORD'),
-        'HOST': env.str('DB_HOST'),
-        'PORT': env.str('DB_PORT'),
-        'OPTIONS': {'sslmode': 'require'},
-    }
-
-LOGGING = {
-    'version': 1,
-    'disable_existing_loggers': False,
-    'handlers': {
-        'console': {
-            'class': 'logging.StreamHandler',
-        },
-    },
-    'root': {
-        'handlers': ['console'],
-        'level': 'WARNING',
-    },
-    'loggers': {
-        'django': {
-            'handlers': ['console'],
-            'level': 'INFO',
-            'propagate': False,
-        },
-    },
+DATABASES['default'] = {
+    'NAME': env.str('DB_NAME'),
+    'ENGINE': 'django.db.backends.postgresql_psycopg2',
+    'USER': env.str('DB_USER'),
+    'PASSWORD': env.str('DB_PASSWORD'),
+    'HOST': env.str('DB_HOST'),
+    'PORT': env.str('DB_PORT'),
 }
 
+# Django Rest Framework Settings
 REST_FRAMEWORK = {
     'DEFAULT_FILTER_BACKENDS': (
         'django_filters.rest_framework.DjangoFilterBackend',
     ),
 }
 
+# Discord bot settings
 DISCORD_URL_BUG = env.str('DISCORD_URL_BUG')
 DISCORD_URL_FEEDBACK = env.str('DISCORD_URL_FEEDBACK')
 
@@ -259,16 +213,3 @@ MESSAGE_TAGS = {
     messages.WARNING: 'alert-warning',
     messages.ERROR: 'alert-danger',
 }
-
-
-def custom_recording_logic(request):
-    """Exclude API views for django-silk"""
-    return not request.path.startswith('/api')
-
-
-# Performance profiling for non-API views during development
-SILKY_INTERCEPT_FUNC = custom_recording_logic
-if DEBUG:
-    INSTALLED_APPS.append('silk')
-    MIDDLEWARE.append('silk.middleware.SilkyMiddleware')
-    SILKY_INTERCEPT_FUNC = custom_recording_logic
